GUI list model perf optimization
Keep SimpleMatchList and UnmatchedObjectListModel sorted
for better lookup performance using binarySearch.
(useful for indexOf() or remove() operations)

Keeping the list sorted has also the advantage
that in case of delete and then Undo
the elements will be reinserted back at the same position.

Add hashMaps to SimpleMatchList for efficient search
by Reference or Subject.

Restore case if primitivesAdded event because of Undo:
If an OsmPrimitive is deleted by the user, the corresponding
conflation case (match, subjectOnly or referenceOnly) is deleted
from the displayed list.
But it was not restored if the OsmPrimitive is re-added because of
Undo. This commit fix this.