Config panel major update to add plenty of options.
Major modification of the configuration panel
to give acces to all the options of the JCS tool suite.
Reused the interface of the original JCS plugin in the JUMP tool
(taken from the OpenJump project).

Added an AttributeMatcher with Levenshtein distance to score tag
value difference.

REM: the new unioning option is able to match many Reference
items to 1 Subject item.
This situation is not handled properly afterward, in particular the
«Conflate» operation will not behave as ones could expect.